+++ app/deployments/tasks.py
@@ -31,6 +31,9 @@ def update_values_for_timeseries(timeseries):
         logger.warning(
             f"Timeout when trying to retrieve dataset {timeseries[0].dataset.name} with constraint {timeseries[0].constraints}: {error}", extra={"timeseries": timeseries, "constraints": timeseries[0].constraints}
         )
+    
+    except OSError as error:
+        logger.error(f"Error loading dataset {timeseries[0].dataset.name} with constraints {timeseries[0].constraints}: {error}", extra={"timeseries": timeseries, "constraints": timeseries[0].constraints})
 
     else:
         for series in timeseries:
@@ -68,6 +71,9 @@ def refresh_dataset(dataset_id: int, healthcheck: bool = False):
     """
     dataset = ErddapDataset.objects.get(pk=dataset_id)
 
+    if healthcheck:
+        dataset.healthcheck_start()
+
     groups = dataset.group_timeseries_by_constraint()
 
     for constraints, timeseries in groups.items():
@@ -86,6 +92,9 @@ def refresh_server(server_id: int, healthcheck: bool = False):
         healthcheck (int): Should Healthchecks.io be singled after all timeseries are updated
     """
     server = ErddapServer.objects.get(pk=server_id)
+
+    if healthcheck:
+        dataset.healthcheck_start()
 
     for ds in server.erddapdataset_set.all():
         refresh_dataset(ds.id)